# docs/swagger.yaml
basePath: /swagger/things-service
definitions:
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  entity.MonitorGatewayDetail:
    properties:
      identifier:
        type: string
      name:
        type: string
      points:
        items:
          $ref: '#/definitions/entity.MonitorGatewayDetailPoint'
        type: array
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
    type: object
  entity.MonitorGatewayDetailPoint:
    properties:
      access_type:
        description: 接入类型 ， IO/485
        type: string
      address_485:
        description: 如果是485，则是地址
        type: string
      address_io:
        description: 如果是io，则是AI1,DI1
        type: string
      device_identifier:
        type: string
      name:
        type: string
      status:
        description: 状态，1：正常，0：异常
        type: integer
      update_time:
        description: 格式 2006-01-02 15:04:05
        type: string
      value:
        description: 值
        type: string
    type: object
  entity.MonitorListDbItem:
    properties:
      cache_hit_rate:
        description: 缓存命中率
        type: number
      conflict_count:
        description: 冲突数
        type: integer
      connection_count:
        description: 连接数
        type: integer
      deadlock_count:
        description: 死锁数
        type: integer
      name:
        description: 名称
        type: string
      server:
        description: 服务器
        type: string
      status:
        description: 运行状态，1：正常，0：停止
        type: integer
    type: object
  entity.MonitorListGatewayItem:
    properties:
      ai_port_count:
        description: ai端口数
        type: integer
      ao_port_count:
        description: ao端口数
        type: integer
      di_port_count:
        description: di端口数
        type: integer
      do_port_count:
        description: do端口数
        type: integer
      identifier:
        description: 标识符
        type: string
      name:
        description: 名称
        type: string
      status:
        description: 运行状态，1：正常，0：停止
        type: integer
    type: object
  entity.MonitorListHostItem:
    properties:
      cpu_usage:
        description: cpu使用率
        type: number
      ip:
        description: ip
        type: string
      kernel:
        type: string
      mem_usage:
        description: 内存使用率
        type: number
      name:
        description: 名称
        type: string
      os:
        type: string
      status:
        description: 状态,1:正常,0:离线
        type: integer
    type: object
  entity.MonitorListInterfaceItem:
    properties:
      alert_count:
        description: 告警次数
        type: integer
      call_frequency:
        description: 调用频率
        type: integer
      latest_data_time:
        type: string
      name:
        description: 名称
        type: string
      protocol:
        description: 接口协议
        type: string
      status:
        description: 状态，1：正常，0：异常
        type: integer
      third_party_system:
        description: 第三方系统
        type: string
    type: object
  entity.MonitorListServiceItem:
    properties:
      host:
        description: 主机
        type: string
      memory_usage:
        description: 占用内存数
        type: integer
      name:
        description: 名称
        type: string
      status:
        description: 运行状态，1：正常，0：停止
        type: integer
      thread_count:
        description: 线程数
        type: integer
    type: object
  entity.MonitorLog:
    properties:
      count:
        type: string
      level:
        type: string
    type: object
  entity.MonitorTotal:
    properties:
      total:
        items:
          $ref: '#/definitions/entity.MonitorTotalItem'
        type: array
    type: object
  entity.MonitorTotalItem:
    properties:
      error:
        description: 异常数
        type: integer
      name:
        description: 名称
        type: string
      normal:
        description: 正常数
        type: integer
      total:
        description: 总数
        type: integer
    type: object
...
  model.User_device:
    properties:
      created_time:
        description: 创建时间
        type: string
      device_id:
        description: 设备id
        type: string
      id:
        description: 唯一标识
        type: string
      index:
        description: 索引排序
        type: integer
      locked:
        description: 是否锁定(0,1)
        type: integer
      updated_time:
        description: 更新时间
        type: string
      user_id:
        description: 用户id
        type: string
    type: object
info:
  contact: {}
  description: things-service API 文档.
  title: things-service RESTful API
  version: "1.0"
paths:
...
  /monitor/gateway/detail:
    get:
      consumes:
      - application/json
      description: 监控接口-gateway-detail
      parameters:
      - description: identifier
        in: query
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.MonitorGatewayDetail'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控接口-gateway-detail
      tags:
      - Monitor
  /monitor/list/dbs:
    get:
      consumes:
      - application/json
      description: 监控数据库列表
      parameters:
      - description: 过滤名称, 不填则为全部
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MonitorListDbItem'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控数据库列表
      tags:
      - Monitor
  /monitor/list/gateways:
    get:
      consumes:
      - application/json
      description: 监控采集网关列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MonitorListGatewayItem'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控采集网关列表
      tags:
      - Monitor
  /monitor/list/hosts:
    get:
      consumes:
      - application/json
      description: 监控主机列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MonitorListHostItem'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控主机列表
      tags:
      - Monitor
  /monitor/list/interfaces:
    get:
      consumes:
      - application/json
      description: 监控接口列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MonitorListInterfaceItem'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控接口列表
      tags:
      - Monitor
  /monitor/list/services:
    get:
      consumes:
      - application/json
      description: 监控服务列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MonitorListServiceItem'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控服务列表
      tags:
      - Monitor
  /monitor/log:
    get:
      consumes:
      - application/json
      description: 监控接口-log
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.MonitorLog'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控接口-log
      tags:
      - Monitor
  /monitor/log/detail:
    get:
      consumes:
      - application/json
      description: 监控接口-log-detail
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 级别， debug,info,error,warn,不填则为全部
        in: query
        name: level
        type: string
      - description: 过滤,可使用正则表达式
        in: query
        name: filter
        type: string
      - description: 限制条数，默认100
        in: query
        name: limit
        type: integer
      - description: 多少小时前，默认24
        in: query
        name: pre_hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.MonitorLog'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控接口-log-detail
      tags:
      - Monitor
  /monitor/log/label/values:
    get:
      consumes:
      - application/json
      description: 监控接口-label values
      parameters:
      - description: label
        in: query
        name: label
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控接口-log-label values
      tags:
      - Monitor
...
  /monitor/total:
    get:
      consumes:
      - application/json
      description: 监控总览
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.MonitorTotal'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 监控总览
      tags:
      - Monitor
...
  /user-device:
    post:
      consumes:
      - application/json
      description: 保存实例
      parameters:
      - description: 实例全部信息
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.User_device'
      produces:
      - application/json
      responses:
        "200":
          description: 实例
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User_device'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 保存实例
      tags:
      - User_device
...
  /user-device/{id}:
    delete:
      description: 删除实例
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User_device'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除实例
      tags:
      - User_device
  /user-device/batch-delete:
    post:
      consumes:
      - application/json
      description: 批量删除
      parameters:
      - description: id数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 批量删除
      tags:
      - User_device
  /user-device/batch-upsert:
    post:
      consumes:
      - application/json
      description: 批量更新
      parameters:
      - description: 对象数组
        in: body
        name: entities
        required: true
        schema:
          items:
            $ref: '#/definitions/model.User_device'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 批量更新
      tags:
      - User_device
swagger: "2.0"